+++ modules/hooks/scene-hooks.js
@@ -1,22 +1,19 @@
-import { HMTokenHooks } from './token-hooks.js';
-
 export class HMSceneControlHooks {
     static getSceneControlButtons(controls) {
-        const tools = controls.find((a) => a.name === 'token').tools;
         const reachControl = {
             icon: 'far fa-circle-dot',
-            name: 'reachview',
-            title: 'CONTROLS.reachview',
+            name: 'hmReach',
+            title: 'HM.CONTROLS.hmReach',
             toggle: true,
             active: false,
-            onClick: (active) => { game.user.showAllThreats = active; },
+            visible: true,
+            button: true,
+            onChange: (_event, active) => {
+                game.user.showAllThreats = active;
+                canvas.tokens.draw();
+            },
         };
-        tools.push(reachControl);
-    }
-
-    static renderSceneControls() {
-        if (!canvas.tokens) return;
-        const placeables = canvas.tokens.placeables.filter((t) => t.visible);
-        if (placeables.length) placeables.forEach((t) => HMTokenHooks.drawToken(t));
+        const { tools } = controls.tokens;
+        tools.hmReach = reachControl;
     }
 }