+++ modules/hooks/support-hooks.js
@@ -1,7 +1,4 @@
-/* global dragRuler */
-import { SYSTEM_ID } from '../tables/constants.js';
 import { diceSoNiceRollStartHandler } from '../integrations/dice-so-nice.js';
-import { dragRulerHandler } from '../integrations/drag-ruler.js';
 import { tokenHPAttribute, HM_ER_SPEED } from '../integrations/elevationruler.js';
 
 export class HMSupportHooks {
@@ -15,10 +12,5 @@ export class HMSupportHooks {
 
     static diceSoNiceRollStart(...args) {
         return diceSoNiceRollStartHandler(...args);
-    }
-
-    static dragRulerReady(...args) {
-        const HMSpeedProvider = dragRulerHandler(...args);
-        dragRuler.registerSystem(SYSTEM_ID, HMSpeedProvider);
     }
 }