+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+## Added
+- v13 support.
+
+## Fixed
+- Title for saves dialog.
+
+## Removed
+- v12 support.
+- dragRuler support.
+
 ## [0.4.25] - 2025-03-04
 ## Added
 - Attack/Defend/Damage macros now remember your last used weapon.